Move useCompletedSessions
